evas: fix massive number of warning introduced by Evas_3d.

+++ src/lib/evas/canvas/evas_3d_camera.c
@@ -44,8 +44,8 @@ evas_3d_camera_node_add(Evas_3D_Camera *camera, Evas_3D_Node *node)
              return;
           }
      }
-   else count = (int)eina_hash_find(pd->nodes, &node);
-   eina_hash_set(pd->nodes, &node, (const void *)(count + 1));
+   else count = (int)(uintptr_t)eina_hash_find(pd->nodes, &node);
+   eina_hash_set(pd->nodes, &node, (const void *)(uintptr_t)(count + 1));
 }
 
 void
@@ -60,9 +60,9 @@ evas_3d_camera_node_del(Evas_3D_Camera *camera, Evas_3D_Node *node)
         return;
      }
 
-   count = (int)eina_hash_find(pd->nodes, &node);
+   count = (int)(uintptr_t)eina_hash_find(pd->nodes, &node);
    if (count == 1) eina_hash_del(pd->nodes, &node, NULL);
-   else eina_hash_set(pd->nodes, &node, (const void *)(count - 1));
+   else eina_hash_set(pd->nodes, &node, (const void *)(uintptr_t)(count - 1));
 }
 
 EOLIAN static void
